Updated README.md
this is not the final product, just wanted to start the README.md file. Have to complete the following: Testing set, Evaluation, Challenges and Learnings. Will update it with the analysis. Feel free to add! 

README(project).md was the original README.md file that I just switched by changing the name a bit